% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefMulti.R
\name{coefMulti}
\alias{coefMulti}
\title{Application d'un coefficient multiplicateur selon la difficulte.}
\usage{
coefMulti(distance)
}
\arguments{
\item{distance}{une difficulte.}
}
\value{
un coefficient en fonction de \code{distance} .
}
\description{
Application d'un coefficient multiplicateur selon la difficulte.
}
\examples{
coefMulti("XXL")
[1] 6
}
