% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mboServiceConnect.R
\name{mboServiceConnect}
\alias{mboServiceConnect}
\title{Connect to a mlrMBO-as-a-service webserver.}
\usage{
mboServiceConnect(hostname)
}
\arguments{
\item{hostname}{The server to connect to. Can start with either "http://" or "https://".}
}
\value{
A handle, which can be passed to the other functions.
}
\description{
Connect to a mlrMBO-as-a-service webserver.
}
