% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mboServiceDisconnect.R
\name{mboServiceDisconnect}
\alias{mboServiceDisconnect}
\title{Disconnect from the mlrMBO-as-a-service webserver.}
\usage{
mboServiceDisconnect(obj)
}
\arguments{
\item{obj}{The handle object created by mboServiceConnect.}
}
\description{
Disconnect from the mlrMBO-as-a-service webserver.
}
