% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mboServicePropose.R
\name{mboServicePropose}
\alias{mboServicePropose}
\title{Request a point proposal from the mlrMBO-as-a-service webserver.}
\usage{
mboServicePropose(obj)
}
\arguments{
\item{obj}{The handle object created by mboServiceConnect.}
}
\value{
A vector containing the X-values of an interesting point.
}
\description{
Request a point proposal from the mlrMBO-as-a-service webserver.
}
