% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mboServiceSetConfigKeyRaw.R
\name{mboServiceSetConfigKeyRaw}
\alias{mboServiceSetConfigKeyRaw}
\title{Change the mlrMBO settings used for your point proposal.
This is a internal function usable for arbitrary keys and values.
Use mboServiceSetConfigKey instead.}
\usage{
mboServiceSetConfigKeyRaw(obj, key, value)
}
\arguments{
\item{obj}{The handle object created by mboServiceConnect.}

\item{key}{The config key to be set.}

\item{value}{The new value of the key.}
}
\description{
Change the mlrMBO settings used for your point proposal.
This is a internal function usable for arbitrary keys and values.
Use mboServiceSetConfigKey instead.
}
