% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mboServiceUpload.R
\name{mboServiceUpload}
\alias{mboServiceUpload}
\title{Uploads your data to the service, such that it can be evaluated.}
\usage{
mboServiceUpload(obj, data)
}
\arguments{
\item{obj}{The handle object created by mboServiceConnect.}

\item{data}{The data as a data.frame(...). The target column should be named y.}
}
\description{
Uploads your data to the service, such that it can be evaluated.
}
