\name{mlrMBOservice-package}
\alias{mlrMBOservice-package}
\alias{mlrMBOservice}
\docType{package}
\title{
\packageTitle{mlrMBOservice}
}
\description{
\packageDescription{mlrMBOservice}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mlrMBOservice}
\packageIndices{mlrMBOservice}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{mlrMBOservice}

Maintainer: \packageMaintainer{mlrMBOservice}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# ~~ simple examples of the most important functions ~~
}
