% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocpu.R
\name{ocpu}
\alias{ocpu}
\alias{ocpu_server}
\alias{ocpu_post_json}
\alias{ocpu_post_encoded}
\alias{ocpu_post_multipart}
\alias{ocpu_post_pb}
\alias{ocpu_post_rds}
\title{OpenCPU API}
\usage{
ocpu(path, handle = new_handle(), server = ocpu_server(),
  stop_on_error = TRUE)

ocpu_server(server = NULL)

ocpu_post_json(path, args = NULL)

ocpu_post_encoded(path, args = NULL)

ocpu_post_multipart(path, args = NULL)

ocpu_post_pb(path, args = NULL)

ocpu_post_rds(path, args = NULL)
}
\arguments{
\item{path}{target api (without the \code{/ocpu} part)}

\item{handle}{a curl handle to pass custom options and headers}

\item{server}{base URL to OpenCPU server}

\item{args}{a named list with function arguments, c.f. \code{\link[=do.call]{do.call()}}}
}
\description{
Calls an R function or script via the OpenCPU protocol and return output.
}
\examples{
# Note that server might send cached responses
ocpu_post_json('/library/stats/R/rnorm', list(n = 5, mean = 3))
ocpu_post_encoded('/library/stats/R/rnorm', list(n = 5, mean = 3))
ocpu_post_multipart('/library/stats/R/rnorm', list(n = 5, mean = 3))
ocpu_post_pb('/library/stats/R/rnorm', list(n = 5, mean = 3))
ocpu_post_rds('/library/stats/R/rnorm', list(n = 5, mean = 3))
}
