% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ocpu_options}
\alias{ocpu_options}
\title{Client options}
\usage{
ocpu_options(...)
}
\arguments{
\item{...}{named curl options passed to \code{\link[curl:handle_setopt]{curl::handle_setopt()}}}
}
\description{
Global \link[curl:curl_options]{options} applied to each request to OpenCPU.
}
\examples{
# Turn on verbose for all requests:
ocpu_options(verbose = TRUE)

# Turn off again
ocpu_options(verbose = NULL)
}
