% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocpu.R
\name{ocpu}
\alias{ocpu}
\alias{ocpu_post_pb}
\alias{ocpu_post_json}
\title{OpenCPU API}
\usage{
ocpu(path, handle = NULL, server = "https://cloud.opencpu.org/ocpu")

ocpu_post_pb(path, args = NULL)

ocpu_post_json(path, args = NULL)
}
\arguments{
\item{path}{target api (without the \code{/ocpu} part)}

\item{handle}{a curl handle to pass custom options and headers}

\item{server}{base URL to OpenCPU}

\item{args}{a named list with function arguments, c.f. \code{\link[=do.call]{do.call()}}}
}
\description{
Calls an R function or script via the OpenCPU protocol and return output.
}
\examples{
# Note server might send cached responses
ocpu_post_pb('/library/stats/R/rnorm', list(n = 5, mean = 3))
ocpu_post_json('/library/stats/R/rnorm', list(n = 5, mean = 3))
}
